+++ src/router/index.tsx
@@ -2,7 +2,7 @@ import AIChat from '@/pages/AIChat'
 import AutoMessage from '@/pages/AutoMessage'
 import AutoPopUp from '@/pages/AutoPopUp'
 import AutoReply from '@/pages/AutoReply'
-import BrowserControl from '@/pages/BrowserControl'
+import LiveControl from '@/pages/LiveControl'
 import Settings from '@/pages/SettingsPage'
 import { createHashRouter } from 'react-router'
 import App from '../App'
@@ -14,7 +14,7 @@ export const router = createHashRouter([
     children: [
       {
         path: '/',
-        element: <BrowserControl />,
+        element: <LiveControl />,
       },
       {
         path: '/auto-message',
